CBL-5429: Logging "correlation-ID" for all logs logged by Replicator.
Also
- Reverted f6b93e28765c81cf6905b5d271c046715d04cf83 to reinstate checking of format string and -Wformat-nonliteral.
- Logging Coll=<collection index> using the same mechanism.
- Corrected several places where collection ID is repeated in the log.

+++ LiteCore/Support/LogEncoder.cc
@@ -249,10 +249,10 @@ namespace litecore {
     }
 
     void LogEncoder::_writeStringToken(const char* token) {
-        const auto name = _formats.find((size_t)token);
+        const auto name = _formats.find(token);
         if ( name == _formats.end() ) {
             const auto n = (unsigned)_formats.size();
-            _formats.insert({(size_t)token, n});
+            _formats.insert({token, n});
             _writeUVarInt(n);
             _writer.write(token, strlen(token) + 1);  // add the actual string the first time
         } else {
